update:move MockRedisClient to test dir
